models were modified 'cause they didn't have a relationship, New views in their directories

+++ database/migrations/2016_11_17_223437_create_alumnos_table.php
@@ -18,7 +18,7 @@ class CreateAlumnosTable extends Migration
             $table->string('nombre', 70);
             $table->string('apellido_paterno', 50);
             $table->string('apellido_materno', 50)->nullable();
-            $table->smallInteger('grupo_id');
+            $table->tinyInteger('grupo_id');
         });
     }
